Prevent orphans submit from leaf nodes from crashing the pool
storeBlock in local_comms.js validates whether it should accept a block
dependent on whether it can find its hash in the blockchain. This is fine
for the master node where it's checking the very blockchain it submit the
solution to.  However, when a leaf node submit a block that turned out to be
an orphan, this logic would cause an infinite loop where storeBlock refused
to advance since it could never find the orphan in it's own blockchain.

This resolves the matter by changing getBlockHeaderByHash() to
getBlockHeaderByHeight().  We only care if we *can* validate by being at least
at the height of the submitted block.  blockManager will orphan the block on its
own later if it truly is an orphan, and we still get a valid block value to store
as the reward.

While this still "loops" until the master node is at the submitted block's height,
we introduce a 2 second delay when responding with a negative response to storeBlock.
In most cases, one or two iterrations will be enough for the master node to be at
the desired height to validate.

Also fixed an issue where the SQL table 'block_log' was not ever having orphans tagged
due to an incorrect attempt to use 'true' instead of '1' for the field.